fix: remove TestResult struct duplication across test files
- Create common test_helper.hpp with TestResult definition
- Remove duplicate TestResult structs from all test files
- Add template and string overload support for assert_eq
- Eliminates 388 lines of code duplication
