+++ tests/test_tsplib.cpp
@@ -7,60 +7,9 @@
 
 #include <evolab/evolab.hpp>
 
+#include "test_helper.hpp"
+
 using namespace evolab::io;
-
-struct TestResult {
-    int passed = 0;
-    int failed = 0;
-
-    void assert_true(bool condition, const std::string& message) {
-        if (condition) {
-            passed++;
-            std::cout << "[PASS] " << message << "\n";
-        } else {
-            failed++;
-            std::cout << "[FAIL] " << message << "\n";
-        }
-    }
-
-    void assert_eq(int expected, int actual, const std::string& message) {
-        assert_true(expected == actual, message + " (expected: " + std::to_string(expected) +
-                                            ", actual: " + std::to_string(actual) + ")");
-    }
-
-    void assert_eq(size_t expected, size_t actual, const std::string& message) {
-        assert_true(expected == actual, message + " (expected: " + std::to_string(expected) +
-                                            ", actual: " + std::to_string(actual) + ")");
-    }
-
-    void assert_eq(double expected, double actual, const std::string& message,
-                   double tolerance = 1e-6) {
-        assert_true(std::abs(expected - actual) < tolerance,
-                    message + " (expected: " + std::to_string(expected) +
-                        ", actual: " + std::to_string(actual) + ")");
-    }
-
-    void assert_eq(const std::string& expected, const std::string& actual,
-                   const std::string& message) {
-        assert_true(expected == actual,
-                    message + " (expected: '" + expected + "', actual: '" + actual + "')");
-    }
-
-    template <typename T>
-    void assert_eq(T expected, T actual, const std::string& message) {
-        assert_true(expected == actual,
-                    message + " (expected: " + std::to_string(static_cast<int>(expected)) +
-                        ", actual: " + std::to_string(static_cast<int>(actual)) + ")");
-    }
-
-    int summary() {
-        std::cout << "\n=== TSPLIB Test Summary ===\n";
-        std::cout << "Passed: " << passed << "\n";
-        std::cout << "Failed: " << failed << "\n";
-        std::cout << "Total:  " << (passed + failed) << "\n";
-        return failed == 0 ? 0 : 1;
-    }
-};
 
 void test_basic_parsing(TestResult& result) {
     std::string tsp_content = R"(
